Moved quality cli arg test to a separate file.
